+++ pkg/operator/event_handler.go
@@ -17,9 +17,22 @@ package operator
 import (
 	"fmt"
 	"log/slog"
+	"strings"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/tools/cache"
 )
+
+type FilterFunc func(interface{}) bool
+
+type EventHandlerOption func(*EventHandler)
+
+// The object is enqueued only if FilterFunc returns true.
+func WithFilter(filter FilterFunc) EventHandlerOption {
+	return func(eh *EventHandler) {
+		eh.filterFuncs = append(eh.filterFuncs, filter)
+	}
+}
 
 // EventHandler implements the k8s.io/tools/cache.ResourceEventHandler interface.
 type EventHandler struct {
@@ -29,7 +42,10 @@ type EventHandler struct {
 
 	objName     string
 	enqueueFunc func(string)
+	filterFuncs []FilterFunc
 }
+
+var _ = cache.ResourceEventHandler(&EventHandler{})
 
 func NewEventHandler(
 	logger *slog.Logger,
@@ -37,23 +53,37 @@ func NewEventHandler(
 	metrics *Metrics,
 	objName string,
 	enqueueFunc func(ns string),
+	options ...EventHandlerOption,
 ) *EventHandler {
-	return &EventHandler{
+	eh := &EventHandler{
 		logger:      logger,
 		accessor:    accessor,
 		metrics:     metrics,
 		objName:     objName,
 		enqueueFunc: enqueueFunc,
 	}
+
+	for _, opt := range options {
+		opt(eh)
+	}
+
+	return eh
 }
 
 func (e *EventHandler) OnAdd(obj interface{}, _ bool) {
 	o, ok := e.accessor.ObjectMetadata(obj)
-	if ok {
-		e.logger.Debug(fmt.Sprintf("%s added", e.objName))
-		e.metrics.TriggerByCounter(e.objName, AddEvent).Inc()
-		e.enqueueFunc(o.GetNamespace())
+	if !ok {
+		return
 	}
+
+	for _, fn := range e.filterFuncs {
+		if !fn(obj) {
+			return
+		}
+	}
+
+	e.recordEvent(AddEvent, o)
+	e.enqueueFunc(o.GetNamespace())
 }
 
 func (e *EventHandler) OnUpdate(old, cur interface{}) {
@@ -61,17 +91,45 @@ func (e *EventHandler) OnUpdate(old, cur interface{}) {
 		return
 	}
 
+	for _, fn := range e.filterFuncs {
+		if !fn(cur) {
+			return
+		}
+	}
+
 	if o, ok := e.accessor.ObjectMetadata(cur); ok {
-		e.logger.Debug(fmt.Sprintf("%s updated", e.objName))
-		e.metrics.TriggerByCounter(e.objName, UpdateEvent)
+		e.recordEvent(UpdateEvent, o)
 		e.enqueueFunc(o.GetNamespace())
 	}
 }
 
 func (e *EventHandler) OnDelete(obj interface{}) {
-	if o, ok := e.accessor.ObjectMetadata(obj); ok {
-		e.logger.Debug(fmt.Sprintf("%s deleted", e.objName))
-		e.metrics.TriggerByCounter(e.objName, DeleteEvent).Inc()
-		e.enqueueFunc(o.GetNamespace())
+	o, ok := e.accessor.ObjectMetadata(obj)
+	if !ok {
+		return
 	}
+
+	for _, fn := range e.filterFuncs {
+		if !fn(obj) {
+			return
+		}
+	}
+
+	e.recordEvent(DeleteEvent, o)
+	e.enqueueFunc(o.GetNamespace())
+}
+
+func (e *EventHandler) recordEvent(event HandlerEvent, obj metav1.Object) {
+	eventName := string(event)
+	if strings.HasSuffix(eventName, "e") {
+		eventName += "d"
+	} else {
+		eventName += "ed"
+	}
+
+	e.logger.Debug(
+		fmt.Sprintf("%s %s", e.objName, eventName),
+		strings.ToLower(e.objName), fmt.Sprintf("%s/%s", obj.GetNamespace(), obj.GetName()),
+	)
+	e.metrics.TriggerByCounter(e.objName, event).Inc()
 }